feat: 신청 로직 구현 및 관리자에게 메시지 보내는 로직 추가
